+++ demoBoot/src/main/java/com/listTest/maxAndMinAndSumDemo.java
@@ -13,18 +13,18 @@ import java.util.List;
 public class maxAndMinAndSumDemo {
     public static void main(String[] args) {
         List<User> list=new ArrayList<>();
-        list.add(new User(1,"lm1","lm11"));
-        list.add(new User(2,"lm2","lm22"));
-        list.add(new User(3,"lm3","lm33"));
-        list.add(new User(4,"lm4","lm44"));
-        list.add(new User(5,"lm5",null));
-        list.add(new User(6,"lm6","lm66"));
-        list.add(new User(7,"lm6","lm77"));
-        Integer id = list.stream().max(Comparator.comparing(User::getId)).get().getId();
+        list.add(new User(1L,"lm1","lm11"));
+        list.add(new User(2L,"lm2","lm22"));
+        list.add(new User(3L,"lm3","lm33"));
+        list.add(new User(4L,"lm4","lm44"));
+        list.add(new User(5L,"lm5",null));
+        list.add(new User(6L,"lm6","lm66"));
+        list.add(new User(7L,"lm6","lm77"));
+        Long id = list.stream().max(Comparator.comparing(User::getId)).get().getId();
         System.out.println(id);
-        Integer id1 = list.stream().min(Comparator.comparing(User::getId)).get().getId();
+        Long id1 = list.stream().min(Comparator.comparing(User::getId)).get().getId();
         System.out.println(id1);
-        Integer sum = list.stream().map(e -> e.getId()).reduce(Integer::sum).get();
+        Long sum = list.stream().map(e -> e.getId()).reduce(Long::sum).get();
         System.out.println(sum);
     }
 }
